Refactor qtree spec tests
The type spec tests should not put any dependencies on providers so
create dummy providers and use them instead of the real ones.

As a sideeffect we have to require puppet/util/network_device at a
puppet type level (instead of provider level) because the libraries
would not get loaded otherwise and `apply_to_device` (used at a type
level) seems to depend on these libraries.