Reload quota if necessary
Modifying a quota has no effect until the next time quota-on or quota-
resize is called. The API call quota-resize is supposed to be faster but
does only work if an existing entry was modified. If we have created a
new quota that was absent before, we'll have to turn quota off and than
turn it back on.

This is now realized in the flush method: flush is called after every
property has been synced and is only called if at least one property
was out of sync: this is exactly what we want.