Add Puppet 2.7.x support for netapp_commands
The netapp_commands method can be used to add class and instances
wrapper methods to invoke the NetApp API calls. The methods were created
with the `create_class_and_instance_method` method that was added in
puppet 3.0.0 and is not available when running puppet 2.7.x

Detect the missing `create_class_and_instances_method` method and use a
work around to archieve the same result. As the result the code now
works for puppet 3.x as well as puppet 2.7.x