Remove the special config file
To execute Puppet for a NetApp device it is necessary to add an entry
to the device configuration file. This device configuration file holds
the certificate name of the device and the type of the device (in this
case netapp)

In the past you had to edit a second `netapp.yml` file where you can drop
the correct user credentials for your NetApp device. This does seem
redundant as the device configuration already specifies an url attribute
that can easily be used to store a username, a password and the hostname
of the device.

Use the URI instead and drop the special configuration file. This does
also make testing easier as you can overwrite the URI with a facts

Example

    FACTER_url=https://admin:secret@pfiler.example.com puppet resource netapp_user