Rename memory and cpu type facts
We should try to map the device attributes as closely to well-known
facter facts as possible. Facter 1.7. has a `memorysize_mb` fact to
reflect the size in MiB as a number with no unit and a `memorysize`
fact to store the value with a proper unit (in this case "MB"). Use
these two facts to hold the memorysize of a NetApp device.

Facter 1.7 does only have a `processor<INDEX>` fact, but no `processor`
fact. But it defines another fact called `hardwareisa` that contains
the processor type.

Use the name `hardwareisa` to hold the current CPU type on a NetApp
devices and drop the `processor` fact.