Fix failing tests in netapp_export_spec
The netapp_export type defines a validation method that can raise an
ArgumentError if there is an invalid combination of property values.

But if such an error is raised puppet will actually handle it, add more
information about the error and then will then raise an
Puppet::ResourceError including the original error message. As a
result the spec tests did fail:

expected ArgumentError with message matching /Readonly and Readwrite
params cannot be the same./, got #<Puppet::ResourceError: Validation of
Netapp_export[/vol/volume] failed: Readonly and Readwrite params cannot
be the same.>

Expect the general Puppet::Error instead an ArgumentError