Handle the fact that net-ifconfig-get is not available
In case the API call net-ifconfig-get is not available the API call
fails. As a result we should not expect to can access any child
elements.

Check the result of the net-ifconfig-get api call and skip the interface
facts if it the result is "failed".

The unit tests are suboptimal here because we don't test the positive
branch in case the api calls succeed. Unfortunately I don't have a
system available where the API calls works so I don't now how such an
answer would look like.