List necessary access roles in README.markdown
If you want to access your NetApp filers with a dedicated user, you'll
have to assign special capabilities in order to be able to use the
netapp module.

NetApp defines certain categories of rules (see
https://library.netapp.com/ecmdocs/ECMP1196890/html/man1/na_useradmin.1.html)

A `login-` capability is needed to log into the filer. Since we access
the filer through the web api, we need the `login-http-admin` capabiliy.

To use an url of `https://username:password@pfiler/vfiler` (accessing a
virtual filer through the physical filer) we optionally need the
`security-api-vfiler` capabiliy.

And then there is one capability per API call we want to send to the
filer.

List all necessary capabilities in the `README.markdown` so a user of
the module is able to create the necessary role to use the module.